#49 Create ImportSample and update import samples with new controls
