+++ FXyz-Samples/src/main/java/org/fxyz3d/samples/importers/ImportMaya.java
@@ -29,25 +29,8 @@
 
 package org.fxyz3d.samples.importers;
 
-import javafx.animation.Timeline;
-import javafx.beans.InvalidationListener;
-import javafx.beans.Observable;
-import javafx.beans.property.IntegerProperty;
-import javafx.beans.property.ObjectProperty;
-import javafx.beans.property.SimpleIntegerProperty;
-import javafx.beans.property.SimpleObjectProperty;
-import javafx.scene.Node;
-import javafx.scene.Parent;
-import javafx.scene.shape.CullFace;
-import javafx.scene.shape.DrawMode;
-import javafx.scene.shape.MeshView;
-import org.fxyz3d.controls.ControlCategory;
-import org.fxyz3d.controls.NumberSliderControl;
-import org.fxyz3d.controls.factory.ControlFactory;
 import org.fxyz3d.importers.Importer3D;
 import org.fxyz3d.importers.Model3D;
-import org.fxyz3d.samples.shapes.ShapeBaseSample;
-import org.fxyz3d.shapes.polygon.PolygonMeshView;
 
 import java.io.IOException;
 import java.util.logging.Level;
@@ -57,11 +40,7 @@ import java.util.logging.Logger;
  *
  * @author JosePereda
  */
-public class ImportMaya extends ShapeBaseSample<Node> {
-
-    private final IntegerProperty subdivision = new SimpleIntegerProperty(this, "Subdivision Level", 0);
-
-    private final ObjectProperty<Timeline> timeline = new SimpleObjectProperty<>(this, "timeline");
+public class ImportMaya extends ImportSample {
 
     public static void main(String[] args){launch(args);}
     
@@ -69,86 +48,10 @@ public class ImportMaya extends ShapeBaseSample<Node> {
     protected void createMesh() {
         try {
             Model3D modelData = Importer3D.loadAsPoly(getClass().getResource("King_WalkCycle.ma"));
-
             model = modelData.getRoot();
-
-             modelData.getTimeline().ifPresent(t -> {
-                timeline.set(t);
-
-                model.sceneProperty().addListener(new InvalidationListener() {
-                    @Override
-                    public void invalidated(Observable observable) {
-                        if (model.getScene() == null) {
-                            model.sceneProperty().removeListener(this);
-                            timeline.set(null);
-                        }
-                    }
-                });
-            });
+            modelData.getTimeline().ifPresent(timeline::set);
         } catch (IOException ex) {
             Logger.getLogger(ImportMaya.class.getName()).log(Level.SEVERE, null, ex);
         }
-    }
-    
-    @Override
-    protected void addMeshAndListeners() {
-        drawMode.addListener((obs, b, b1) -> {
-            if (model != null) {
-                setDrawMode(model, b1);
-            }
-        });
-        culling.addListener((obs, b, b1) -> {
-            if (model != null) {
-                setCullFace(model, b1);
-            }
-        });
-        subdivision.addListener((obs, ov, nv) -> setSubdivisionLevel(model, nv.intValue()));
-    }
-    
-    private void setDrawMode(Node node, DrawMode value) {
-        if (node instanceof PolygonMeshView) {
-            ((PolygonMeshView) node).setDrawMode(value);
-        } else if (node instanceof MeshView) {
-            ((MeshView) node).setDrawMode(value);
-        } else if (node instanceof Parent) {
-            ((Parent) node).getChildrenUnmodifiable().forEach(n -> setDrawMode(n, value));
-        }
-    }
-    
-    private void setCullFace(Node node, CullFace value) {
-        if (node instanceof PolygonMeshView) {
-            ((PolygonMeshView) node).setCullFace(value);
-        } else if (node instanceof MeshView) {
-            ((MeshView) node).setCullFace(value);
-        } else if (node instanceof Parent) {
-            ((Parent) node).getChildrenUnmodifiable().forEach(n -> setCullFace(n, value));
-        }
-    }
-    
-    private void setSubdivisionLevel(Node node, int subdivisionLevel) {
-        if (node instanceof PolygonMeshView) {
-            ((PolygonMeshView) node).setSubdivisionLevel(subdivisionLevel);
-        } else if (node instanceof Parent) {
-            ((Parent) node).getChildrenUnmodifiable().forEach(n -> setSubdivisionLevel(n, subdivisionLevel));
-        }
-    }
-    
-    @Override
-    protected Node buildControlPanel() {
-        NumberSliderControl subdivisionSlider = ControlFactory.buildNumberSlider(subdivision, 0, 2);
-        subdivisionSlider.getSlider().setMajorTickUnit(1);
-        subdivisionSlider.getSlider().setMinorTickCount(0);
-        subdivisionSlider.getSlider().setBlockIncrement(1);
-        subdivisionSlider.getSlider().setSnapToTicks(true);
-        ControlCategory geomControls = ControlFactory.buildCategory("Geometry");
-        geomControls.addControls(subdivisionSlider);
-        
-        this.controlPanel = ControlFactory.buildControlPanel(
-                ControlFactory.buildMeshViewCategory(
-                        this.drawMode,
-                        this.culling
-                ), geomControls,
-                ControlFactory.buildAnimationCategory(timeline));
-        return controlPanel;
     }
 }
