+++ FXyz-Samples/src/main/java/org/fxyz3d/samples/importers/ImportOBJ.java
@@ -29,33 +29,15 @@
 
 package org.fxyz3d.samples.importers;
 
-import javafx.scene.Group;
-import javafx.scene.Node;
-import javafx.scene.shape.MeshView;
-import static javafx.application.Application.launch;
-import org.fxyz3d.controls.factory.ControlFactory;
 import org.fxyz3d.importers.Importer3D;
-import org.fxyz3d.samples.shapes.ShapeBaseSample;
-import org.fxyz3d.shapes.polygon.PolygonMeshView;
 
 import java.io.IOException;
-import java.util.logging.Level;
-import java.util.logging.Logger;
-import javafx.beans.property.IntegerProperty;
-import javafx.beans.property.SimpleIntegerProperty;
-import javafx.scene.Parent;
-import org.fxyz3d.controls.ControlCategory;
-import org.fxyz3d.controls.NumberSliderControl;
 
 /**
  *
  * @author JosePereda
  */
-public class ImportOBJ extends ShapeBaseSample<Group> {
-
-    private final IntegerProperty subdivision = new SimpleIntegerProperty(this, "Subdivision Level", 0);
-    
-    private boolean asPolygonMesh = true;
+public class ImportOBJ extends ImportSample {
 
     public static void main(String[] args){launch(args);}
     
@@ -63,69 +45,9 @@ public class ImportOBJ extends ShapeBaseSample<Group> {
     protected void createMesh() {
         try {
             model = Importer3D.loadAsPoly(getClass().getResource("DukeKing.obj")).getRoot();
-        } catch (IOException ex) {
-            Logger.getLogger(ImportOBJ.class.getName()).log(Level.SEVERE, null, ex);
+        } catch (IOException e) {
+            e.printStackTrace();
         }
     }
-    
-    @Override
-    protected void addMeshAndListeners() {
-        drawMode.addListener((obs, b, b1) -> {
-            if (model != null) {
-                if (asPolygonMesh) {
-                    model.getChildren().stream()
-                            .filter(PolygonMeshView.class::isInstance)
-                            .map(PolygonMeshView.class::cast)
-                            .forEach(pmv -> pmv.setDrawMode(b1));
-                } else {
-                    model.getChildren().stream()
-                            .filter(MeshView.class::isInstance)
-                            .map(MeshView.class::cast)
-                            .forEach(pmv -> pmv.setDrawMode(b1));
-                }
-            }
-        });
-        culling.addListener((obs, b, b1) -> {
-            if (model != null) {
-                if (asPolygonMesh) {
-                    model.getChildren().stream()
-                            .filter(PolygonMeshView.class::isInstance)
-                            .map(PolygonMeshView.class::cast)
-                            .forEach(pmv -> pmv.setCullFace(b1));
-                } else {
-                    model.getChildren().stream()
-                            .filter(MeshView.class::isInstance)
-                            .map(MeshView.class::cast)
-                            .forEach(pmv -> pmv.setCullFace(b1));
-                }
-            }
-        });
-        subdivision.addListener((obs, ov, nv) -> setSubdivisionLevel(model, nv.intValue()));
-    }
-    
-    private void setSubdivisionLevel(Node node, int subdivisionLevel) {
-        if (node instanceof PolygonMeshView) {
-            ((PolygonMeshView) node).setSubdivisionLevel(subdivisionLevel);
-        } else if (node instanceof Parent) {
-            ((Parent) node).getChildrenUnmodifiable().forEach(n -> setSubdivisionLevel(n, subdivisionLevel));
-        }
-    }
-    
-    @Override
-    protected Node buildControlPanel() {
-        NumberSliderControl subdivisionSlider = ControlFactory.buildNumberSlider(subdivision, 0, 2);
-        subdivisionSlider.getSlider().setMajorTickUnit(1);
-        subdivisionSlider.getSlider().setMinorTickCount(0);
-        subdivisionSlider.getSlider().setBlockIncrement(1);
-        subdivisionSlider.getSlider().setSnapToTicks(true);
-        ControlCategory geomControls = ControlFactory.buildCategory("Geometry");
-        geomControls.addControls(subdivisionSlider);
-        
-        this.controlPanel = ControlFactory.buildControlPanel(
-                ControlFactory.buildMeshViewCategory(
-                        this.drawMode,
-                        this.culling
-                ), geomControls);
-        return controlPanel;
-    }
+
 }
